pyproject.toml
[tool.poetry]
name = "ango-xml-parser"
version = "1.0.0"
description = "Auto.ru Парсер"
authors = ["Евгений Клюев <kluev.evga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
django-rest-framework = "^0.1.0"
djangorestframework-xml = "^2.0.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
